+++ PlayWithMax 1.0/CarouselAndGame/Views/CardView.swift
